refactor: Kafka Consumer manual commit 적용 및 파티션 키 방식 변경
- auto commit -> manual commit 전환으로 메시지 손실 방지
  auto commit 사용 시 문제점
  1. 메시지 수신됨
  2. 아직 처리 안 했는데 auto commit으로 offset 저장됨
  3. 처리 중 오류 발생 -> 메시지 실패
  4. 하지만 kafka는 이미 처리된 것으로 간주 -> 메시지 손실 !

  즉, 메시지를 제대로 처리하지도 않았는데 "처리 완료" 표시가 되는 상황 발생

  해결책: 매뉴얼 커밋 사용 시
  1. 메시지 수신됨
  2. 처리 시도 (DB 저장 등)
  3. 처리 성공 시 ack.acknowledge() -> 그때 offset 커밋
  4. 실패 시 ack 생략 -> 메시지는 다시 소비

- topic key: inside.coupon.{couponId} -> inside.coupon + partition key: couponId
