+++ src/main/java/kr/hhplus/be/server/interfaces/config/KafkaConfig.java
@@ -1,0 +1,35 @@
+package kr.hhplus.be.server.interfaces.config;
+
+import org.apache.kafka.clients.admin.AdminClient;
+import org.apache.kafka.clients.admin.AdminClientConfig;
+import org.springframework.boot.autoconfigure.kafka.KafkaProperties;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.kafka.config.ConcurrentKafkaListenerContainerFactory;
+import org.springframework.kafka.core.ConsumerFactory;
+import org.springframework.kafka.listener.ContainerProperties;
+
+import java.util.Map;
+
+@Configuration
+public class KafkaConfig {
+
+    @Bean
+    public AdminClient adminClient(KafkaProperties kafkaProperties) {
+        Map<String, Object> config = Map.of(
+                AdminClientConfig.BOOTSTRAP_SERVERS_CONFIG,
+                String.join(",", kafkaProperties.getBootstrapServers())
+        );
+        return AdminClient.create(config);
+    }
+
+    @Bean
+    public ConcurrentKafkaListenerContainerFactory<String, String> manualAckKafkaListenerContainerFactory(
+            ConsumerFactory<String, String> consumerFactory) {
+        ConcurrentKafkaListenerContainerFactory<String, String> factory =
+                new ConcurrentKafkaListenerContainerFactory<>();
+        factory.setConsumerFactory(consumerFactory);
+        factory.getContainerProperties().setAckMode(ContainerProperties.AckMode.MANUAL_IMMEDIATE);
+        return factory;
+    }
+}
